composebox_typeahead: Suggest dm user in topic recipient typeahead.
This commit includes user suggestion(users with DMs with the current
user or participants in the currently visible conversation view)
for dm in stream topic recipient box typeahead. They show up at the
end of the typeahead list. The right side of the DM user item in the
typeahead list denotes `DM` same as the `NEW` for the new topic item.
Also, when the typeahead list only contains User suggestions then the
footer note is also changed.

Fixes: zulip#31804.